+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVSJ3CN1MF9H1TV8TPZT07KA"
+        "01KVST9YJB5KG0DRNNJN9W2Q1Q"
     )
     r"""The unique identifier of the data source"""