+++ packages/orq-rc/RELEASES.md
@@ -8919,3 +8919,13 @@ Based on:
 - [python v4.11.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.41 - packages/orq-rc
+
+## 2026-06-23 08:44:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.1 (2.911.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.42 - packages/orq-rc